Update bafa_scrapper.py
the new line helps to match the team name for cases wherein the database with the postcodes the team name has an extra character, this issue came from 'Shropshire Revolutions', name in the postcode database, the name in the division 'Shropshire Revolution'. The new line makes up for those extra characters. The assumption is that it refers to the same team. It is unlikely that there are different teams just differentiated by an extra character